gb: maintain permanent lists of useful-files
This is the first step in abandoning the huge memcached process which
serves rpm headers (the cache still can be useful on a smaller scale,
as an LRU cache proper).  Header lists can be somehow "patched", or,
in other words, genpkglist can use its previous output as a primary
cache.  This is already a work in progress.

To get rid of the huge memcached process, though, all other major
scans of the repo should be eliminated as well.  And so this change
introduces a technique of "patching" useful-files lists, instead
of creating them from scratch all over again.  With this change,
rpmquery(1) is executed only for the packages found in plan/add-bin.

This change also introduces an optimization of not running genbasedir
at all, provided that the merged-down arch+noarch list of useful files
hasn't changed.  And another optimization of running genbasedir with
the --maybe-unchanged option, provided that there are no *-debuginfo
subpackages among the arch packages.  This can save an expensive xz
compression step for pkglist.debuginfo.  About 15% arch packages
in the repo do not have *-debuginfo subpackages.

$ cd Sisyphus/files/x86_64/RPMS
$ rpm -qp --qf '%{SOURCERPM}\n' *.rpm |sort -u |wc -l
9080
$ rpm -qp --qf '%{SOURCERPM}\n' *-debuginfo-*.rpm |sort -u |wc -l
7685

So this change is actually beneficial on its own.

Another scan which should be eliminated is the one which is used to
generate contents_index files.  This can only be done after bloated
indices are generated along with the stripped-down indices, i.e. after
the appropriate changes to genbasedir; rpmquery then can be replaced
with pkglist-query on the bloated indices.